requirements.txt
numpy
urllib3<2.0
scikit-learn
tensorflow
tensorflow-macos
tensorflow-metal
psutil
pillow
matplotlib
opencv-python
faiss-cpu
imblearn
realesrgan
torchvision==0.20.0 